Graph View: Moved ErrorType class to mav section.
Added SetFromSqlRecord method and made members private to improove decoupling.
Extended Documentation
Removed the Timestamp from the != operator.
Inverted error logic in "AP2DataPlot2DModel::getErrorValues()"
