+++ src/ui/AP2DataPlot2DModel.h
@@ -32,30 +32,8 @@ This file is part of the APM_PLANNER project
 
 #include <QAbstractTableModel>
 #include <QSqlDatabase>
+#include <ArduPilotMegaMAV.h>
 
-
-//**************** Helper Class ErrorType *****************
-
-/**
- * @brief Helper class for making it easier to handle the
- *        errorcodes, especially the toString conversion
- */
-
-class ErrorType
-{
-public:
-    u_int64_t Timeus;
-    u_int8_t SubSys;
-    u_int8_t ErrorCode;
-
-    ErrorType();
-
-    bool operator != (const ErrorType &lhs);
-
-    QString toString();
-};
-
-// ********************************************************
 
 class AP2DataPlot2DModel : public QAbstractTableModel
 {
